+++ packages/core/src/components/tree-view/tree-view.stories.ts
@@ -1,8 +1,9 @@
 import { storiesOfTreeView } from '@dso-toolkit/sources';
 import { storiesOf } from '@storybook/web-components';
 
-import { treeViewTemplate } from './tree-view.template';
 import readme from './readme.md';
+import { html } from 'lit-html';
+import { treeViewTemplate } from './tree-view.template';
 
 storiesOfTreeView(
   {
@@ -11,6 +12,16 @@ storiesOfTreeView(
     readme
   },
   {
-    treeViewTemplate
+    treeViewDemoTemplate: (collection, onOpenItem, onCloseItem) => html`
+      ${treeViewTemplate({
+        collection,
+        onOpenItem: function (e) {
+          onOpenItem(this.collection, e.detail, collection => this.collection = collection);
+        },
+        onCloseItem: function (e) {
+          onCloseItem(this.collection, e.detail, collection => this.collection = collection);
+        }
+      })}
+    `
   }
 );
