+++ packages/core/src/components/tree-view/tree-view.tsx
@@ -1,63 +1,7 @@
-import { Component, h, ComponentInterface, Prop, Event, EventEmitter, Element, FunctionalComponent } from '@stencil/core';
-import clsx from 'clsx';
+import { Component, h, ComponentInterface, Prop, Event, EventEmitter } from '@stencil/core';
 
-import { TreeViewItem, TreeViewItemIcon } from '@dso-toolkit/sources';
-
-interface TreeViewItemProps<T> {
-  owner: TreeView;
-  ancestors: TreeViewItem<T>[];
-  item: TreeViewItem<T>;
-  index: number;
-  level: number;
-  setSize: number;
-}
-
-const DsoTreeItem: FunctionalComponent<TreeViewItemProps<string>> = ({ owner, ancestors, item, index, level, setSize }) => (
-  <li
-    key={item.reference}
-    class={clsx('tree-item', { 'has-child': item.hasItems }, { 'collapsed': !item.open })}
-  >
-    <div class="tree-branch-control">
-      {item.hasItems
-        ?
-          <div onClick={(e) => owner.itemClick(e, ancestors, item)}>
-            <dso-icon icon={item.open ? 'chevron-down' : 'chevron-right'}></dso-icon>
-          </div>
-        : <dso-icon></dso-icon>
-      }
-    </div>
-    <p
-      class="tree-content"
-      tabindex="0"
-      role="treeitem"
-      aria-expanded={item.hasItems ? '' + !!item.open : undefined}
-      aria-level={level}
-      aria-setsize={setSize}
-      aria-posinset={index + 1}
-    >
-      <a role="link">{item.label}</a>
-      {item.icons?.map((icon: TreeViewItemIcon) =>
-        <dso-icon icon={icon.icon} title={icon.label}></dso-icon>
-      )}
-    </p>
-    <ul role="group">
-      {item.hasItems && !item.items && item.loading
-        ? <dso-progress-indicator size="small" label="Resultaten laden: een moment geduld alstublieft." />
-        : undefined
-      }
-      {item.items?.map((childItem: TreeViewItem<string>, index: number, org: TreeViewItem<string>[]) =>
-        <DsoTreeItem
-          owner={owner}
-          ancestors={[...ancestors, item]}
-          item={childItem}
-          index={index}
-          level={level + 1}
-          setSize={org.length}
-        ></DsoTreeItem>
-      )}
-    </ul>
-  </li>
-);
+import { TreeViewItem } from '@dso-toolkit/sources';
+import { DsoTreeItem } from './tree-item';
 
 @Component({
   tag: 'dso-tree-view',
@@ -68,20 +12,11 @@ export class TreeView implements ComponentInterface {
   @Prop()
   collection!: TreeViewItem<string>[];
 
-  @Prop()
-  onFetchItems!: Function;
+  @Event()
+  openItem!: EventEmitter<TreeViewItem<string>[]>;
 
   @Event()
-  toggleItem!: EventEmitter<TreeViewItem<string>[]>;
-
-  @Event()
-  fetchItems!: EventEmitter<TreeViewItem<string>[]>;
-
-  @Element()
-  ref: any;
-
-  @Prop()
-  updateItems = TreeView.updateDeepTree;
+  closeItem!: EventEmitter<TreeViewItem<string>[]>;
 
   keyDownListener = (event: KeyboardEvent) => {
     if (event.defaultPrevented) {
@@ -134,14 +69,11 @@ export class TreeView implements ComponentInterface {
       return;
     }
 
-    const expandableItem = event.target.closest('li');
-
-    if (item.hasItems && !item.items) {
-      expandableItem?.classList.remove('collapsed');
-      this.fetchItems.emit([...ancestors, item]);
+    if (item.open) {
+      this.closeItem.emit([...ancestors, item]);
     }
     else {
-      this.toggleItem.emit([...ancestors, item]);
+      this.openItem.emit([...ancestors, item]);
     }
   }
 
@@ -204,33 +136,6 @@ export class TreeView implements ComponentInterface {
     if (controlElement instanceof HTMLElement) {
       controlElement.click();
     }
-  }
-
-  private static updateDeepTree = function(
-    collection: TreeViewItem<string>[],
-    path: TreeViewItem<string>[],
-    updater: (item: TreeViewItem<string>) => TreeViewItem<string>
-    ): TreeViewItem<string>[] {
-    const item = collection.find(item => item.reference === path[0].reference);
-
-    return collection.map(root => {
-      if (root !== item) {
-        return root;
-      }
-
-      if (path.length === 1) {
-        return updater(item);
-      }
-
-      if (root === item) {
-        return {
-          ...item,
-          items: item.items ? TreeView.updateDeepTree(item.items, path.slice(1), updater) : item.items
-        };
-      }
-
-      throw new Error('Updating items collection encountered an error.');
-    });
   }
 
   render() {
