--- packages/sources/src/components/tree-view/tree-view.args.ts
@@ -1,7 +1,0 @@
-import { ArgTypes } from '../../stories-helpers';
-
-export interface TreeViewArgs {
-}
-
-export const treeViewArgTypes: ArgTypes<TreeViewArgs> = {
-};
