+++ packages/sources/src/components/tree-view/tree-view.stories.ts
@@ -1,12 +1,14 @@
-import { ArgsStoryFn } from '@storybook/addons';
 import { StorybookParameters } from '../../stories-helpers';
 
-import { treeViewArgTypes } from './tree-view.args';
-import { items, subItems } from './tree-view.content';
-import { TreeView } from './tree-view.models';
+import { TreeViewDemo } from './tree-view.demo';
+import { TreeViewItem } from './tree-view.models';
 
 export interface TreeViewParameters<TemplateFnReturnType> {
-  treeViewTemplate: (treeViewProperties: TreeView<string>) => TemplateFnReturnType;
+  treeViewDemoTemplate: (
+    collection: TreeViewItem<string>[],
+    onOpenItem: (tree: TreeViewItem<string>[], path: TreeViewItem<string>[], callback: (collection: TreeViewItem<string>[]) => void) => void,
+    onCloseItem: (tree: TreeViewItem<string>[], path: TreeViewItem<string>[], callback: (collection: TreeViewItem<string>[]) => void) => void
+  ) => TemplateFnReturnType;
 }
 
 export function storiesOfTreeView<TemplateFnReturnType>(
@@ -16,27 +18,21 @@ export function storiesOfTreeView<TemplateFnReturnType>(
     readme
   }: StorybookParameters,
   {
-    treeViewTemplate,
+    treeViewDemoTemplate,
   }: TreeViewParameters<TemplateFnReturnType>
 ) {
   const stories = storiesOf('Tree View', mainModule)
     .addParameters({
+      controls: {
+        hideNoControlsWarning: true
+      },
       docs: {
         page: readme
-      },
-      argTypes: treeViewArgTypes
+      }
     });
 
   stories.add(
     'Tree View',
-    (args: any) => treeViewTemplate(args),
-    {
-      args: {
-        collection: items,
-        onFetchItems: (event: any) => subItems,
-        onItemOpen: (event: any) => items[0] = { ...items[0], open: true },
-        onItemClose: (event: any) => items[0] = { ...items[0], open: false },
-      }
-    }
+    () => treeViewDemoTemplate(TreeViewDemo.collection, TreeViewDemo.onOpenItem, TreeViewDemo.onCloseItem)
   );
 }
